Increase functionality
